AIs can only see within their view distance
This makes it possible for the player to defeat an AI orbiter! The player can guess the location of the enemy ship based on when they last saw it. The player can also see the edge of the enemy ship before the enemy can see them (the enemy must be able to see the centre of the player's ship).